refactor: Group recommendation view tools.
