+++ src/tools/index.ts
@@ -33,12 +33,12 @@ import "./list-costs";
 import "./list-dashboards";
 import "./list-folders";
 import "./list-provider-resources";
-import "./list-recommendation-views";
 import "./list-recommendations";
 import "./list-tag-values";
 import "./list-tags";
 import "./list-unit-costs";
 import "./query-costs";
+import "./recommendation-views";
 import "./submit-user-feedback";
 import "./update-anomaly";
 import "./update-folder";
