+++ src/tools/recommendation-views/index.ts
@@ -1,0 +1,1 @@
+import "./list-recommendation-views";
