+++ src/tools/recommendation-views/list-recommendation-views.test.ts
@@ -1,7 +1,6 @@
 import type { GetRecommendationViewsResponse } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
-import tool from "./list-recommendation-views";
-import { DEFAULT_LIMIT } from "./structure/constants";
+import { DEFAULT_LIMIT } from "../structure/constants";
 import {
   type ExecutionTestTableItem,
   type ExtractOutputSchema,
@@ -10,7 +9,8 @@ import {
   requestsInOrder,
   type SchemaTestTableItem,
   testTool,
-} from "./utils/testing";
+} from "../utils/testing";
+import tool from "./list-recommendation-views";
 
 type Validators = ExtractValidators<typeof tool>;
 type OutputSchema = ExtractOutputSchema<typeof tool>;
